Enhancements and cleanup

+++ src/main.cpp
@@ -150,10 +150,10 @@ int main(int argc, char** argv)
 	case GDT_UInt16:
 		dem_data = new uint16_t[size];
 		break;
-	default: 
+	default:
 		ERR << "Unexpected DEM band data type";
 		exit(1);
-		
+
 	}
 
 	if (dem_band->RasterIO(GF_Read, 0, 0, w, h, dem_data, w, h, dem_band_type, 0, 0) != CE_None) {
@@ -197,10 +197,10 @@ int main(int argc, char** argv)
 		switch (dem_band_type) {
 		case GDT_Float32:
 			process_image<float>(image_path, out_path, ProcessingParameters<float> {
-					params.skip_visibility_test,
+				params.skip_visibility_test,
 					shot,
 					has_nodata,
-					static_cast<float>(no_data),
+					no_data,
 					transform,
 					static_cast<double>(dem_offset_x),
 					static_cast<double>(dem_offset_y),
@@ -211,16 +211,16 @@ int main(int argc, char** argv)
 					static_cast<float*>(dem_data),
 					params.interpolation,
 					params.with_alpha,
-					wkt				
-				}
+					wkt
+			}
 			);
 			break;
 		case GDT_Byte:
 			process_image<uint8_t>(image_path, out_path, ProcessingParameters<uint8_t> {
-					params.skip_visibility_test,
+				params.skip_visibility_test,
 					shot,
 					has_nodata,
-					static_cast<float>(no_data),
+					no_data,
 					transform,
 					static_cast<double>(dem_offset_x),
 					static_cast<double>(dem_offset_y),
@@ -231,16 +231,16 @@ int main(int argc, char** argv)
 					static_cast<uint8_t*>(dem_data),
 					params.interpolation,
 					params.with_alpha,
-					wkt					
-				}
+					wkt
+			}
 			);
 			break;
 		case GDT_UInt16:
 			process_image<uint16_t>(image_path, out_path, ProcessingParameters<uint16_t> {
-					params.skip_visibility_test,
+				params.skip_visibility_test,
 					shot,
 					has_nodata,
-					static_cast<float>(no_data),
+					no_data,
 					transform,
 					static_cast<double>(dem_offset_x),
 					static_cast<double>(dem_offset_y),
@@ -251,8 +251,8 @@ int main(int argc, char** argv)
 					static_cast<uint16_t*>(dem_data),
 					params.interpolation,
 					params.with_alpha,
-					tmp_wkt					
-				}
+					tmp_wkt
+			}
 			);
 			break;
 		default:
